+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - (Feature) Add Feature dependency
 - (Feature) Run secured containers as a feature
 - (Feature) Expose core.PodSecurityContext Sysctl options
+- (Bugfix) Skip Collection check for missing Database
 
 ## [1.2.31](https://github.com/arangodb/kube-arangodb/tree/1.2.31) (2023-07-14)
 - (Improvement) Block traffic on the services if there is more than 1 active leader in ActiveFailover mode